Move parser test code to an integration test
