+++ src/parser/mod.rs
@@ -17,4 +17,3 @@ mod util;
 
 pub mod compiled;
 pub mod expansion;
-pub mod source;
