+++ tests/parser.rs
@@ -1,0 +1,327 @@
+//            DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
+//                    Version 2, December 2004
+//
+// Copyleft (ↄ) meh. <meh@schizofreni.co> | http://meh.schizofreni.co
+//
+// Everyone is permitted to copy and distribute verbatim or modified
+// copies of this license document, and changing it is allowed as long
+// as the name is changed.
+//
+//            DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
+//   TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION
+//
+//  0. You just DO WHAT THE FUCK YOU WANT TO.
+
+use nom::branch::alt;
+use nom::bytes::streaming::{tag, take, take_until, take_while};
+use nom::character::streaming::char;
+use nom::character::{is_digit, streaming::line_ending as eol};
+use nom::combinator::{complete, eof, map, map_res, opt};
+use nom::error::{make_error, ErrorKind};
+use nom::sequence::terminated;
+use nom::IResult;
+use std::borrow::Cow;
+use std::fs::File;
+use std::io::Read;
+use std::str;
+
+const NONE: u8 = 0b000000;
+const PRINT: u8 = 0b000001;
+const SPACE: u8 = 0b000010;
+const CONTROL: u8 = 0b000100;
+const PIPE: u8 = 0b001000;
+const COMMA: u8 = 0b010000;
+const EOL: u8 = 0b100000;
+
+// Ugly table of DOOM, gotta run and gun.
+#[rustfmt::skip]
+static ASCII: [u8; 256] = [
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, SPACE, EOL, NONE, NONE, EOL, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	PRINT | SPACE, PRINT, PRINT, PRINT | CONTROL, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT | COMMA | CONTROL, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT | CONTROL, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT, PRINT,
+	PRINT, PRINT, PRINT, PRINT, PRINT | PIPE, PRINT, PRINT, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+	NONE, NONE, NONE, NONE, NONE, NONE, NONE, NONE,
+];
+
+#[inline(always)]
+pub fn is_ws(ch: u8) -> bool {
+	unsafe { ASCII.get_unchecked(ch as usize) & SPACE == SPACE }
+}
+
+#[inline(always)]
+pub fn is_eol(ch: u8) -> bool {
+	unsafe { ASCII.get_unchecked(ch as usize) & EOL == EOL }
+}
+
+#[inline(always)]
+pub fn is_printable_no_pipe(ch: u8) -> bool {
+	unsafe { ASCII.get_unchecked(ch as usize) & (PRINT | PIPE) == PRINT }
+}
+
+#[inline(always)]
+pub fn is_printable_no_comma(ch: u8) -> bool {
+	unsafe { ASCII.get_unchecked(ch as usize) & (PRINT | COMMA) == PRINT }
+}
+
+#[inline(always)]
+pub fn is_printable_no_control(ch: u8) -> bool {
+	unsafe { ASCII.get_unchecked(ch as usize) & (PRINT | CONTROL) == PRINT }
+}
+
+pub fn ws(input: &[u8]) -> IResult<&[u8], char> {
+	alt((char(' '), char('\t')))(input)
+}
+
+pub fn end(input: &[u8]) -> IResult<&[u8], &[u8]> {
+	alt((eof, eol))(input)
+}
+
+pub fn unescape(i: &[u8]) -> Cow<'_, [u8]> {
+	fn escape<I: Iterator<Item = u8>>(output: &mut Vec<u8>, iter: &mut I) {
+		match iter.next() {
+			None => (),
+
+			Some(b'a') => output.push(0x07),
+
+			Some(b'b') => output.push(0x08),
+
+			Some(b'E') | Some(b'e') => output.push(0x1B),
+
+			Some(b'f') => output.push(0x0C),
+
+			Some(b'l') | Some(b'n') => output.push(b'\n'),
+
+			Some(b'r') => output.push(b'\r'),
+
+			Some(b's') => output.push(b' '),
+
+			Some(b't') => output.push(b'\t'),
+
+			Some(b'^') => output.push(b'^'),
+
+			Some(b'\\') => output.push(b'\\'),
+
+			Some(b',') => output.push(b','),
+
+			Some(b':') => output.push(b':'),
+
+			Some(b'0') => output.push(0x00),
+
+			Some(a) if is_digit(a) => match (iter.next(), iter.next()) {
+				(Some(b), Some(c)) if is_digit(b) && is_digit(c) => {
+					if let Ok(number) =
+						u8::from_str_radix(unsafe { str::from_utf8_unchecked(&[a, b, c]) }, 8)
+					{
+						output.push(number);
+					} else {
+						output.extend(&[a, b, c]);
+					}
+				}
+
+				(Some(b), None) => output.extend(&[b'\\', a, b]),
+
+				(None, None) => output.extend(&[b'\\', a]),
+
+				_ => unreachable!(),
+			},
+
+			Some(ch) => output.extend(&[b'\\', ch]),
+		}
+	}
+
+	fn control<I: Iterator<Item = u8>>(output: &mut Vec<u8>, iter: &mut I) {
+		match iter.next() {
+			None => (),
+
+			Some(ch) if ch.is_ascii_uppercase() => output.push(ch - b'A' + 1),
+
+			Some(ch) if ch.is_ascii_lowercase() => output.push(ch - b'a' + 1),
+
+			Some(ch) => output.extend(&[b'^', ch]),
+		}
+	}
+
+	let mut chars = i.iter().cloned();
+	let mut offset = 0;
+
+	while let Some(ch) = chars.next() {
+		if ch == b'\\' || ch == b'^' {
+			let mut output = i[..offset].to_vec();
+
+			match ch {
+				b'\\' => escape(&mut output, &mut chars),
+
+				b'^' => control(&mut output, &mut chars),
+
+				_ => unreachable!(),
+			}
+
+			while let Some(ch) = chars.next() {
+				match ch {
+					b'\\' => escape(&mut output, &mut chars),
+
+					b'^' => control(&mut output, &mut chars),
+
+					ch => output.push(ch),
+				}
+			}
+
+			return Cow::Owned(output);
+		}
+
+		offset += 1;
+	}
+
+	Cow::Borrowed(i)
+}
+
+#[derive(Eq, PartialEq, Clone, Debug)]
+pub enum Item<'a> {
+	Comment(&'a str),
+
+	Definition { name: &'a str, aliases: Vec<&'a str>, description: &'a str },
+
+	True(&'a str),
+	Number(&'a str, i16),
+	String(&'a str, Cow<'a, [u8]>),
+	Disable(&'a str),
+}
+
+pub fn parse(input: &[u8]) -> IResult<&[u8], Item<'_>> {
+	alt((comment, definition, disable, entry))(input)
+}
+
+fn comment(input: &[u8]) -> IResult<&[u8], Item<'_>> {
+	let (input, _) = tag("#")(input)?;
+	let (input, content) = map_res(terminated(take_until("\n"), tag("\n")), str::from_utf8)(input)?;
+	let (input, _) = opt(complete(take_while(is_eol)))(input)?;
+
+	Ok((input, Item::Comment(content.trim())))
+}
+
+fn definition(input: &[u8]) -> IResult<&[u8], Item<'_>> {
+	let (input, name) =
+		map(take_while(is_printable_no_pipe), |n| unsafe { str::from_utf8_unchecked(n) })(input)?;
+
+	let (input, _) = tag("|")(input)?;
+
+	let (input, content) =
+		map(take_while(is_printable_no_comma), |n| unsafe { str::from_utf8_unchecked(n) })(input)?;
+
+	let (input, _) = tag(",")(input)?;
+
+	let (input, _) = take_while(is_ws)(input)?;
+
+	let (input, _) = eol(input)?;
+	let (input, _) = opt(complete(take_while(is_eol)))(input)?;
+
+	Ok((input, {
+		let mut aliases = content.split('|').map(|n| n.trim()).collect::<Vec<_>>();
+
+		Item::Definition { name, description: aliases.pop().unwrap(), aliases }
+	}))
+}
+
+fn disable(input: &[u8]) -> IResult<&[u8], Item<'_>> {
+	let (input, _) = ws(input)?;
+	let (input, _) = take_while(is_ws)(input)?;
+	let (input, _) = tag("@")(input)?;
+
+	let (input, name) =
+		map(take_while(is_printable_no_control), |n| unsafe { str::from_utf8_unchecked(n) })(
+			input,
+		)?;
+
+	let (input, _) = tag(",")(input)?;
+	let (input, _) = take_while(is_ws)(input)?;
+	let (input, _) = end(input)?;
+	let (input, _) = opt(complete(take_while(is_eol)))(input)?;
+
+	Ok((input, Item::Disable(name)))
+}
+
+fn entry(input: &[u8]) -> IResult<&[u8], Item<'_>> {
+	let (input, _) = ws(input)?;
+	let (input, _) = take_while(is_ws)(input)?;
+
+	let (input, name) =
+		map(take_while(is_printable_no_control), |n| unsafe { str::from_utf8_unchecked(n) })(
+			input,
+		)?;
+
+	let (input, c) = take(1_usize)(input)?;
+	let (input, value) = match c {
+		b"," => (input, Item::True(name)),
+
+		b"#" => {
+			let (input, value) =
+				map(take_while(is_digit), |n| unsafe { str::from_utf8_unchecked(n) })(input)?;
+
+			let (input, _) = tag(",")(input)?;
+
+			(input, Item::Number(name, value.parse().unwrap()))
+		}
+
+		b"=" => {
+			let (input, value) = take_while(is_printable_no_comma)(input)?;
+
+			let (input, _) = tag(",")(input)?;
+
+			(input, Item::String(name, unescape(value)))
+		}
+
+		_ => Err(nom::Err::Error(make_error(input, ErrorKind::Switch)))?,
+	};
+
+	let (input, _) = take_while(is_ws)(input)?;
+	let (input, _) = end(input)?;
+	let (input, _) = opt(complete(take_while(is_eol)))(input)?;
+
+	Ok((input, value))
+}
+
+#[test]
+fn parsing() {
+	let mut file = File::open("tests/xterm.terminfo").unwrap();
+	let mut buffer = Vec::new();
+	file.read_to_end(&mut buffer).unwrap();
+
+	let mut input = &buffer[..];
+
+	while !input.is_empty() {
+		match parse(input) {
+			Ok((rest, _)) => input = rest,
+
+			Err(::nom::Err::Incomplete(_)) => panic!("incomplete"),
+
+			Err(err) => panic!("parsing: {:?}", err),
+		}
+	}
+}
